Massive cleanup of the entire codebase.  Notable changes include:
- Use templates to slash code duplication in primitive-integer constructors and
  converters.
- Remove excessively chatty and historical comments, and move some pieces of
  design discussion to the proper places.
- Improve code format.

Not everything is perfect, but this is much better, and the conversion problem
is finally fixed.  I've gotten to the point where the sample program runs
without crashing and produces the correct output.  More refinements to come.

+++ sample.cc
@@ -18,8 +18,8 @@ int main() {
 		BigInteger a; // a is 0
 		int b = 535;
 
-		a = b; // From int to BigInteger...
-		b = a; // ...and back, no casts required!
+		a = b; // From int to BigInteger implicitly...
+		b = a.toInt(); // ...and back explicitly.
 		/*
 		 * If a were too big for an int you'd get a runtime exception.
 		 * The Big Integer Library throws C-strings (that is,
@@ -119,4 +119,4 @@ Running the sample program produces this output:
 314^9 = 29673367320587092457984
 314^10 = 9317437338664347031806976
 
- */
+*/
